Finished

+++ SensitivityArea.py
@@ -3,8 +3,8 @@ from dataclasses import dataclass
 
 @dataclass
 class SensitivityArea:
-    x_min: float
-    x_max: float
-    y_min: float
-    y_max: float
-    min_size: float
+    x_min: int
+    x_max: int
+    y_min: int
+    y_max: int
+    min_size: int
